IDisposable and 64bit

+++ TweetWatch/NativeMethods.cs
@@ -1,16 +1,12 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Runtime.InteropServices;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace TweetWatch
 {
     internal static class NativeMethods
     {
         [DllImport("user32.dll")]
-        public static extern int SendMessage(IntPtr handle, int msg, int wParam, int lParam);
+        public static extern IntPtr SendMessage(IntPtr handle, int msg, IntPtr wParam, IntPtr lParam);
 
         [DllImport("user32.dll")]
         public static extern int ReleaseCapture();
